Create interal wrapper for ErrorHandlerBoundary
